dotnet core refactor
