+++ AmbientContext.sln
@@ -1,0 +1,36 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 14
+VisualStudioVersion = 14.0.25420.1
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{8310E2CB-EC96-436B-94FA-7C6E515C7E28}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "test", "test", "{8D2E0840-4EF5-40C9-9B48-874FEC4E2350}"
+EndProject
+Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "AmbientContext", "src\AmbientContext\AmbientContext.xproj", "{65C8B019-078D-4E62-80AB-A668E5BBD0BA}"
+EndProject
+Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "AmbientContext.Tests", "test\AmbientContext.Tests\AmbientContext.Tests.xproj", "{6DDE60C0-D993-4C6D-AC0D-FBABB77DB49B}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{65C8B019-078D-4E62-80AB-A668E5BBD0BA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{65C8B019-078D-4E62-80AB-A668E5BBD0BA}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{65C8B019-078D-4E62-80AB-A668E5BBD0BA}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{65C8B019-078D-4E62-80AB-A668E5BBD0BA}.Release|Any CPU.Build.0 = Release|Any CPU
+		{6DDE60C0-D993-4C6D-AC0D-FBABB77DB49B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{6DDE60C0-D993-4C6D-AC0D-FBABB77DB49B}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{6DDE60C0-D993-4C6D-AC0D-FBABB77DB49B}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{6DDE60C0-D993-4C6D-AC0D-FBABB77DB49B}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(NestedProjects) = preSolution
+		{65C8B019-078D-4E62-80AB-A668E5BBD0BA} = {8310E2CB-EC96-436B-94FA-7C6E515C7E28}
+		{6DDE60C0-D993-4C6D-AC0D-FBABB77DB49B} = {8D2E0840-4EF5-40C9-9B48-874FEC4E2350}
+	EndGlobalSection
+EndGlobal
