+++ appveyor.yml
@@ -1,44 +1,22 @@
-﻿version: 0.0.1.{build}
-
-environment:
-  packageVersion: 0.0.1
-
-init:
-- ps: $env:buildVersion = "$env:packageVersion.$env:appveyor_build_number"
-- ps: $env:nugetVersion = "$env:packageVersion-beta-$env:appveyor_build_number"
-- ps: Update-AppveyorBuild -Version $env:buildVersion
-
-assembly_info:
-  patch: true
-  file: '**\AssemblyInfo.*'
-  assembly_version: '$(buildVersion)'
-  assembly_file_version: '$(buildVersion)'
-  assembly_informational_version: '$(nugetVersion)'
-
-pull_requests:
+version: '{build}'  
+pull_requests:  
   do_not_increment_build_number: true
-
-branches:
+branches:  
   only:
   - master
-  - release
-
-configuration: Release
-
-nuget:
-  account_feed: true
+nuget:  
   disable_publish_on_pr: true
-
-cache:
-- ./src/packages -> **\packages.config
-
-before_build:
-- nuget restore ./src/AmbientContext.sln
-
-build:
-  project: ./src/AmbientContext.sln
-  publish_nuget: true
-  verbosity: minimal
-
-artifacts:
-- path: NuGet/Release/AmbientContext*.nupkg
+build_script:  
+- ps: .\Build.ps1
+test: off  
+artifacts:  
+- path: .\artifacts\**\*.nupkg
+  name: NuGet
+deploy:
+- provider: NuGet
+  name: production
+  api_key:
+    secure: 'api_key_goes_here'
+  on:
+    branch: master
+    appveyor_repo_tag: true
