+++ src/AmbientContext/project.json
@@ -1,0 +1,28 @@
+{
+  "version": "0.0.2-beta-*",
+  "description": "Ambient context core project.",
+  "authors": [ "Nathan Johnstone" ],
+  "packOptions": {
+    "tags": [ "serilog logging ambient context" ],
+    "projectUrl": "https://github.com/nrjohnstone/AmbientContext",
+    "licenseUrl": "https://github.com/nrjohnstone/AmbientContext/blob/master/LICENSE"
+  },
+
+  "frameworks": {
+    "net451": {
+      "buildOptions": { "define": [ "NET" ] },
+      "dependencies": {
+      }
+    },
+    "netstandard1.6": {
+      "buildOptions": { "define": [ "COREFX" ] },
+      "imports": "dotnet5.4",
+      "dependencies": {
+      }
+    }
+  },
+
+  "dependencies": {
+    "FluentAssertions": "4.14.0" 
+  }
+}
