+++ test/AmbientContext.Tests/project.json
@@ -1,0 +1,33 @@
+{
+  "version": "1.0.0",
+  "description": "Ambientexts.Tests Class Library",
+
+  "testRunner": "xunit",
+
+  "frameworks": {
+    "net451": {
+      "dependencies": {
+        "Microsoft.NETCore.Platforms": "1.0.1"
+      }
+    }
+  },
+  "netcoreapp1.0": {
+    "dependencies": {
+      "Microsoft.NETCore.App": {
+        "type": "platform",
+        "version": "1.0.0"
+      },
+      "System.Linq": "4.1.0"
+    },
+    "imports": [
+      "dnxcore50",
+      "portable-net45+win8"
+    ]
+  },
+
+  "dependencies": {
+    "xunit": "2.2.0-beta2-build3300",
+    "dotnet-test-xunit": "2.2.0-preview2-build1029",
+    "AmbientContext": "0.0.2-beta-*"
+  }
+}
